ride-compare-bot/requirements.txt
python-telegram-bot==21.0
geopy==2.4.1
python-dotenv==1.0.0
